refactor: moved placeholders.ts to meetings folder ml-207
