--- apps/frontend/src/assets/img/meetings/placeholders/placeholders.ts
@@ -1,15 +1,0 @@
-import placeholderBlue from "./placeholder-blue.svg";
-import placeholderGreen from "./placeholder-green.svg";
-import placeholderOrange from "./placeholder-orange.svg";
-import placeholderPurple from "./placeholder-purple.svg";
-import placeholderTeal from "./placeholder-teal.svg";
-
-const PLACEHOLDER_IMAGES = [
-	placeholderBlue,
-	placeholderGreen,
-	placeholderOrange,
-	placeholderPurple,
-	placeholderTeal,
-];
-
-export { PLACEHOLDER_IMAGES };
