+++ apps/frontend/src/modules/meeting/placeholders.ts
@@ -1,0 +1,15 @@
+import placeholderBlue from "~/assets/img/meetings/placeholders/placeholder-blue.svg";
+import placeholderGreen from "~/assets/img/meetings/placeholders/placeholder-green.svg";
+import placeholderOrange from "~/assets/img/meetings/placeholders/placeholder-orange.svg";
+import placeholderPurple from "~/assets/img/meetings/placeholders/placeholder-purple.svg";
+import placeholderTeal from "~/assets/img/meetings/placeholders/placeholder-teal.svg";
+
+const PLACEHOLDER_IMAGES = [
+	placeholderBlue,
+	placeholderGreen,
+	placeholderOrange,
+	placeholderPurple,
+	placeholderTeal,
+];
+
+export { PLACEHOLDER_IMAGES };
